Implemente el calculo de Coeficiente de Produccion del pedido y el calculo de la duracion por fase. Nada de vista todavia, lo probe con tests. Mañana sigo con la reserva de Maquinas.

+++ grails-app/domain/sieym/Maquina.groovy
@@ -5,6 +5,6 @@ class Maquina {
 	String descripcion
 	Fase fase
 	int capacidad
-	int rendimiento
+	float rendimiento
 	int temperatura
 }
